+++ packages/gatsby-transformer-cloudinary/gatsby-node/create-asset-nodes-from-data.js
@@ -11,7 +11,7 @@ exports.createAssetNodesFromData = ({
   createContentDigest,
 }) => {
   const assetDataPaths = getAssetDataPaths({ node });
-  assetDataPaths.forEach(assetDataPath => {
+  assetDataPaths.forEach((assetDataPath) => {
     const assetData = {
       ...get(node, assetDataPath),
     };
@@ -44,7 +44,7 @@ function verifyAssetData(assetData) {
 }
 
 function getAssetDataKeys(node) {
-  return Object.keys(node).filter(key => {
+  return Object.keys(node).filter((key) => {
     return node[key] && node[key].cloudinaryAssetData === true;
   });
 }
@@ -53,22 +53,22 @@ function getAssetDataPaths({ node, basePath = '' }) {
   const currentNode = basePath === '' ? node : get(node, basePath);
 
   const directAssetDataPaths = Object.keys(currentNode)
-    .filter(key => {
+    .filter((key) => {
       return currentNode[key] && currentNode[key].cloudinaryAssetData === true;
     })
-    .map(subPath => {
+    .map((subPath) => {
       return basePath === '' ? subPath : `${basePath}.${subPath}`;
     });
 
   const objectPaths = Object.keys(currentNode)
-    .filter(key => {
+    .filter((key) => {
       return isObject(currentNode[key]);
     })
-    .map(subPath => {
+    .map((subPath) => {
       return basePath === '' ? subPath : `${basePath}.${subPath}`;
     });
 
-  const indirectAssetDataPaths = flatMap(objectPaths, objectPath => {
+  const indirectAssetDataPaths = flatMap(objectPaths, (objectPath) => {
     return getAssetDataPaths({ node, basePath: objectPath });
   });