+++ packages/gatsby-transformer-cloudinary/get-image-objects/get-shared-image-data.js
@@ -36,7 +36,7 @@ exports.getImageURL = ({
 
 // Retrieve aspect ratio if in transformation else create aspect ratio values
 exports.getAspectRatio = (transformations, originalAspectRatio) => {
-  const arTransform = transformations.find(t => t.startsWith('ar_'));
+  const arTransform = transformations.find((t) => t.startsWith('ar_'));
   if (!arTransform) {
     return originalAspectRatio;
   }
@@ -99,35 +99,35 @@ function logBase64Retrieval(url, reporter) {
   if (typeof reporter.info === 'function') {
     reporter.info(
       `[gatsby-transformer-cloudinary] Fetching base64 image ` +
-        `#${fetchedBase64ImageCount} from Cloudinary: ${url}`,
+        `#${fetchedBase64ImageCount} from Cloudinary: ${url}`
     );
     if (fetchedBase64ImageCount == 100) {
       reporter.info(
-        '[gatsby-transformer-cloudinary] ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~',
+        '[gatsby-transformer-cloudinary] ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~'
       );
       reporter.info(
-        'It looks like your project has a lot of images. To improve your build times, ',
+        'It looks like your project has a lot of images. To improve your build times, '
       );
       reporter.info(
-        'you should consider (1) using images that are already on Cloudinary, (2) ',
+        'you should consider (1) using images that are already on Cloudinary, (2) '
       );
       reporter.info(
-        'precomputing the base64 images and providing them to this plugin as the ',
+        'precomputing the base64 images and providing them to this plugin as the '
       );
       reporter.info(
-        '`defaultBase64` property, and (3) setting the plugin option ',
+        '`defaultBase64` property, and (3) setting the plugin option '
       );
       reporter.info(
-        '`alwaysUseDefaultBase64` to true. Doing so will reduce the number of base64 ',
+        '`alwaysUseDefaultBase64` to true. Doing so will reduce the number of base64 '
       );
       reporter.info(
-        'images that need to be fetched from Cloudinary and speed up your Gatsby "query" ',
+        'images that need to be fetched from Cloudinary and speed up your Gatsby "query" '
       );
       reporter.info(
-        'build steps. (See the section "Use images already on Cloudinary" in the README.)',
+        'build steps. (See the section "Use images already on Cloudinary" in the README.)'
       );
       reporter.info(
-        '[gatsby-transformer-cloudinary] ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~',
+        '[gatsby-transformer-cloudinary] ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~'
       );
     }
   }